pyproject.toml
[tool.poetry]
name = "langchain-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.36.0"
gradio = "^4.38.1"
openai = "^1.36.0"
tiktoken = "^0.7.0"
langchain-core = "^0.2.22"
langchain-community = "^0.2.9"
langchain-openai = "^0.1.17"
anthropic = "^0.31.2"
langchain-anthropic = "^0.1.20"
langchain-google-genai = "^1.0.7"
google-generativeai = "^0.7.2"
pymupdf = "^1.24.9"
faiss-cpu = "^1.8.0.post1"
pypdf2 = "^3.0.1"
langsmith = "^0.1.94"
duckduckgo-search = "^6.2.5"
html2text = "^2024.2.26"
lxml = {extras = ["html-clean"], version = "^5.2.2"}
readability-lxml = "^0.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
